+++ src/tools/getBuildFailureLogs/inputSchema.ts
@@ -1,6 +1,22 @@
 import { z } from 'zod';
+import {
+  branchDescription,
+  projectSlugDescription,
+} from '../sharedInputSchemas.js';
 
 export const getBuildFailureOutputInputSchema = z.object({
+  projectSlug: z.string().describe(projectSlugDescription).optional(),
+  branch: z.string().describe(branchDescription).optional(),
+  projectURL: z
+    .string()
+    .describe(
+      'The URL of the CircleCI project. Can be any of these formats:\n' +
+        '- Project URL with branch: https://app.circleci.com/pipelines/gh/organization/project?branch=feature-branch\n' +
+        '- Pipeline URL: https://app.circleci.com/pipelines/gh/organization/project/123\n' +
+        '- Workflow URL: https://app.circleci.com/pipelines/gh/organization/project/123/workflows/abc-def\n' +
+        '- Job URL: https://app.circleci.com/pipelines/gh/organization/project/123/workflows/abc-def/jobs/xyz',
+    )
+    .optional(),
   workspaceRoot: z
     .string()
     .describe(
@@ -14,24 +30,6 @@ export const getBuildFailureOutputInputSchema = z.object({
     .describe(
       'The URL of the remote git repository. This should be the URL of the repository that you cloned to your local workspace. ' +
         'For example: "https://github.com/user/my-project.git"',
-    )
-    .optional(),
-  branch: z
-    .string()
-    .describe(
-      'The name of the branch currently checked out in local workspace. ' +
-        'This should match local git branch. ' +
-        'For example: "feature/my-branch", "bugfix/123", "main", "master" etc.',
-    )
-    .optional(),
-  projectURL: z
-    .string()
-    .describe(
-      'The URL of the CircleCI project. Can be any of these formats:\n' +
-        '- Project URL with branch: https://app.circleci.com/pipelines/gh/organization/project?branch=feature-branch\n' +
-        '- Pipeline URL: https://app.circleci.com/pipelines/gh/organization/project/123\n' +
-        '- Workflow URL: https://app.circleci.com/pipelines/gh/organization/project/123/workflows/abc-def\n' +
-        '- Job URL: https://app.circleci.com/pipelines/gh/organization/project/123/workflows/abc-def/jobs/xyz',
     )
     .optional(),
 });